Move tests that GET the `/metrics` endpoint to the platform router
Tests that make assertions about how metrics are incremented remain in
`test_metrics.py`
